bot.py
from telegram.ext import ApplicationBuilder, CommandHandler, CallbackQueryHandler
from handlers import * 
from config import * 


if __name__ == "__main__":
    app = ApplicationBuilder().token(TELEGRAM_TOKEN).build()
    app.add_handler(CommandHandler("start", start))
    app.add_handler(CallbackQueryHandler(button_handler))

    print("Бот запущен...")
    app.run_polling()

